Add a function for picking rock,paper or scissors
This new update adds a function that allows the computer
to randomly pick rock, paper or scissors.

An array storing this options was created, and a random
number based on the length of the array was picked.

This number was then used as an indexing variable to
select one of the three available choices that were
stored in the array.